tests/wpt/tests/css/css-anchor-position/align-dialog-001.html

<!DOCTYPE html>
<title>Tests the dialog keyword is parsed and computed as specified</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-1/#conditional-centering">
<link rel="author" href="mailto:tabatkins@google.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<div id="container">
  <div id="target"></div>
</div>

<script>
test_valid_value('align-self', 'dialog');
test_valid_value('align-items', 'dialog');
test_valid_value('justify-self', 'dialog');
test_valid_value('justify-items', 'dialog');

test_computed_value('align-self', 'dialog');
test_computed_value('align-items', 'dialog');
test_computed_value('justify-self', 'dialog');
test_computed_value('justify-items', 'dialog');
</script>
